refactor 1: people module
